Cleaned up code used for teleporting.
