+++ TODO.md
@@ -1,6 +1,10 @@
 # TODO
 
-General things that aren't tracked in TODO comments.
+As best as possible, these should also e tracked in TODO comments.
+
+-   Improve the playing of the teleportation sound so that it plays for everyone in the vicinity (outgoing and incoming).
+
+--------------
 
 -   Change to always show current players as well as those who are offline, ut have saved portals.
 -   I probably need some state management for the menu being displayed so I can handle "back" type of workflows.
@@ -8,6 +12,7 @@ General things that aren't tracked in TODO comments.
     -   Icons for all the buttons.
     -   Better PortalPal texture.
     -   Better behavior and texture pack artwork.
--   Fix the sound so that it plays for everyone in the vicinity of the teleport (outgoing and incoming).
 -   Do I need to escape characters going into the json?
 -   Code review and clean up.
+-   Maybe get better success and failure sounds?
+-   Add JSDoc support to everything.
